update check if

+++ el7-gcc-updater.sh
@@ -76,6 +76,7 @@ else
     ./configure --prefix=$DES/gmp-6.1.2 &>/dev/null
     make -j $JN &>/dev/null && make install &>/dev/null
     cd ..
+    rm -rf gmp-6.1.2*
 fi
 
 # Check if mpfr-3.1.6 is already installed
@@ -90,19 +91,25 @@ else
     ./configure --prefix=$DES/mpfr-3.1.6 --with-gmp=$DES/gmp-6.1.2
     make -j $JN && make install
     cd ..
+    rm -rf mpfr-3.1.6*
 fi
 
-# Install mpc 1.0.3
-rm -rf mpc-1.0.3.tar.gz
-wget https://ftp.gnu.org/gnu/mpc/mpc-1.0.3.tar.gz --no-check-certificate
-tar xvf mpc-1.0.3.tar.gz
-cd mpc-1.0.3
-./configure --prefix=$DES/mpc-1.0.3 --with-gmp=$DES/gmp-6.1.2 --with-mpfr=$DES/mpfr-3.1.6
-make -j $JN && make install
-cd ..
 
-# Clean gmp, mpc and mpfr files and directory
-rm -rf gmp-6.1.2* mpc-1.0.3* mpfr-3.1.6*
+# Check if mpc-1.0.3 is already installed
+if [ -d "$DES/mpc-1.0.3" ]; then
+    echo "mpc-1.0.3 is already installed in $DES. Skipping installation."
+else
+    # Install mpc 1.0.3
+    rm -rf mpc-1.0.3.tar.gz
+    wget https://ftp.gnu.org/gnu/mpc/mpc-1.0.3.tar.gz --no-check-certificate
+    tar xvf mpc-1.0.3.tar.gz
+    cd mpc-1.0.3
+    ./configure --prefix=$DES/mpc-1.0.3 --with-gmp=$DES/gmp-6.1.2 --with-mpfr=$DES/mpfr-3.1.6
+    make -j $JN && make install
+    cd ..
+    rm -rf mpc-1.0.3*
+fi
+
 
 # ld configure
 echo "$DES/gmp-6.1.2/lib"  >> /etc/ld.so.conf
@@ -110,13 +117,20 @@ echo "$DES/mpfr-3.1.6/lib" >> /etc/ld.so.conf
 echo "$DES/mpc-1.0.3/lib"  >> /etc/ld.so.conf
 ldconfig -v
 
-# Install GCC
-rm -rf gcc-$GCC_VERS.tar.xz
-wget https://ftp.gnu.org/gnu/gcc/gcc-$GCC_VERS/gcc-$GCC_VERS.tar.xz
-tar xvf gcc-$GCC_VERS.tar.xz
-cd gcc-$GCC_VERS
-./configure --enable-checking=release --enable-languages=c,c++,fortran --disable-multilib \
+# Check if gcc is already installed
+if [ -d "$DES/gcc-$GCC_VERS" ]; then
+    echo "gcc-$GCC_VERS is already installed in $DES. Skipping installation."
+else
+    # Install GCC
+    rm -rf gcc-$GCC_VERS.tar.xz
+    wget https://ftp.gnu.org/gnu/gcc/gcc-$GCC_VERS/gcc-$GCC_VERS.tar.xz
+    tar xvf gcc-$GCC_VERS.tar.xz
+    cd gcc-$GCC_VERS
+    ./configure --enable-checking=release --enable-languages=c,c++,fortran --disable-multilib \
                                                 --prefix=$DES/gcc-$GCC_VERS --with-gmp=$DES/gmp-6.1.2 \
                                                 --with-mpfr=$DES/mpfr-3.1.6 --with-mpc=$DES/mpc-1.0.3
-make -j $JN
-make install
+    make -j $JN
+    make install
+    cd ..
+    rm -rf gcc-$GCC_VERS/ gcc-$GCC_VERS.tar.xz
+fi
